passing around config like it's going out of style

+++ lotek/lib/highlight.py
@@ -8,49 +8,45 @@ from pygments.formatters import HtmlFormatter
 from pygments.lexers import get_lexer_by_name
 from pygments.util import ClassNotFound
 from lotek.lib.logger import log
-from lotek.lib.context import update_formatter
 _FENCE = re.compile(r"^```(\w*)\n([\s\S]*?)^```[ \t]*$", re.MULTILINE)
+
+highlight_formatter = None
 
 def init_formatter(dirs, config):
     global highlight_formatter
-    from lotek.lib.context import highlight_formatter
+    style = config.features.code_theme
+    last_code_file = dirs.LOTEK / "last_code_theme"
+    last_file = ""
+    
     if highlight_formatter is not None:
         log.debug("highlight formatter already set")
         return
-    style = config.features.code_theme
-    update_formatter(HtmlFormatter(classes=True, style=style))
-    from lotek.lib.context import highlight_formatter as hf
-    last_code_file = dirs.LOTEK / "last_code_theme"
-    last_file = ""
+    
     if last_code_file.exists():
         last_file = last_code_file.read_text()
         if last_file == style:
-            log.info("leaving pygments theme unmodified")
-            # assume unchanged and return
+            log.debug("pygments theme unchanged")
             return
-    log.info("new theme detected: updating to %s (was %s)", style, last_file)
+    
     # theme has changed, update the last code theme
     last_code_file.write_text(style)
     # preserve any existing pygments.css
     backup_filename = (
         dirs.STATIC / f"pygments-backup-{date.today().strftime('%Y%m%d-%H%M%S')}.css"
     )
-    shutil.move(dirs.STATIC / "pygments.css", backup_filename)
-    log.info("backed up existing pygments.css to %s", backup_filename)
+    if (dirs.STATIC / "pygments.css").exists():
+        shutil.move(dirs.STATIC / "pygments.css", backup_filename)
+        log.info("backed up existing pygments.css to %s", backup_filename)
     # write out the new theme's css
+    highlight_formatter = HtmlFormatter(style=style)
     (dirs.STATIC / "pygments.css").write_text(
-        hf.get_style_defs("div.highlight")
+        highlight_formatter.get_style_defs("div.highlight")
     )
 
-def process_code_blocks(dirs, text):
+def process_code_blocks(dirs, config, text):
     global highlight_formatter
-    from lotek.lib.context import highlight_formatter
     if not highlight_formatter:
-        from lotek.lib.context import config
         init_formatter(dirs, config)
-        from lotek.lib.context import highlight_formatter as hf
-        highlight_formatter = hf
-
     def replace(m):
         lang = m.group(1).strip().lower() or "text"
         code = m.group(2)
